feat: refactor interview handling; implement sendInterview and handleSubmit functions, and update imports in InterviewChat component
