+++ app/(tabs)/(interview)/[interviewId].tsx
@@ -7,6 +7,8 @@ import { useTheme } from "@/context/theme-context";
 import { Message } from "@/types/message";
 import { Question } from "@/types/question";
 import { handleFeedback } from "@/utils/interview";
+import { handleSubmit } from "@/utils/interview-submit";
+import { sendInterview } from "@/utils/send-interview";
 import { useAuth } from "@clerk/clerk-expo";
 import { LinearGradient } from "expo-linear-gradient";
 import { useLocalSearchParams } from "expo-router";
@@ -22,7 +24,6 @@ import {
 	TouchableWithoutFeedback,
 	View,
 } from "react-native";
-import uuid from "react-native-uuid";
 
 export default function InterviewChatDynamic() {
 	const [messages, setMessages] = useState<Message[]>([]);
@@ -45,24 +46,15 @@ export default function InterviewChatDynamic() {
 	}, [feedback]);
 
 	useEffect(() => {
-		const sendInterview = async () => {
-			const token = await getToken();
-			await fetch(
-				`${process.env.EXPO_PUBLIC_BASE_URL}/interview/${interviewId}`,
-				{
-					method: "POST",
-					headers: {
-						"Content-Type": "application/json",
-						Authorization: `Bearer ${token}`,
-					},
-					body: JSON.stringify({
-						messages: messages,
-						questionNumber: question?.questionNumber,
-					}),
-				}
+		const sendInterviewData = async () => {
+			await sendInterview(
+				getToken,
+				interviewId.toString(),
+				messages,
+				question!
 			);
 		};
-		sendInterview();
+		sendInterviewData();
 	}, [messages]);
 
 	useEffect(() => {
@@ -87,7 +79,6 @@ export default function InterviewChatDynamic() {
 			}
 
 			const data = await response.json();
-			console.log(JSON.stringify(data, null, 2));
 			setMessages(data.interview.messages || []);
 			setQuestion(data.questionDetails);
 			setSolution(data.solution.solution || "");
@@ -100,96 +91,17 @@ export default function InterviewChatDynamic() {
 		fetchData();
 	}, []);
 
-	const handleSubmit = async () => {
-		if (!input.trim() || isLoading) return;
-
-		const userMessage: Message = {
-			id: uuid.v4(),
-			role: "user",
-			content: input.trim(),
-		};
-
-		setMessages((prev) => [...prev!, userMessage]);
-		setInput("");
-		setIsLoading(true);
-
-		const tempMessageId = uuid.v4();
-		setMessages((prev) => [
-			...prev!,
-			{
-				id: tempMessageId,
-				role: "assistant",
-				content: "",
-			},
-		]);
-
-		try {
-			const token = await getToken();
-			const response = await fetch(
-				`${process.env.EXPO_PUBLIC_BASE_URL}/openai/stream`,
-				{
-					method: "POST",
-					headers: {
-						"Content-Type": "application/json",
-						Authorization: `Bearer ${token}`,
-					},
-					body: JSON.stringify({
-						messages: [...messages, userMessage].map(
-							({ role, content }) => ({
-								role,
-								content,
-							})
-						),
-						solution,
-					}),
-				}
-			);
-
-			if (!response.ok) {
-				console.log(response);
-				const errorData = await response.json();
-				alert(errorData.error || "Failed to fetch response");
-				throw new Error(
-					errorData.message || "Failed to fetch response"
-				);
-			}
-
-			const data = await response.json();
-
-			setMessages((prev) =>
-				prev?.map((message) =>
-					message.id === tempMessageId
-						? { ...message, content: data.message }
-						: message
-				)
-			);
-
-			// const reader = response.body.getReader();
-			// const decoder = new TextDecoder();
-			// let content = "";
-
-			// while (true) {
-			// 	const { done, value } = await reader.read();
-			// 	if (done) break;
-
-			// 	const chunk = decoder.decode(value);
-			// 	content += chunk;
-			// 	console.log(chunk);
-
-			// 	// Update the temporary message with the accumulated content
-			// 	setMessages((prev) =>
-			// 		prev.map((message) =>
-			// 			message.id === tempMessageId
-			// 				? { ...message, content }
-			// 				: message
-			// 		)
-			// 	);
-			// }
-		} catch (error) {
-			console.error("Error:", error);
-		} finally {
-			setIsLoading(false);
-		}
+	const onSubmit = async () => {
+		await handleSubmit({
+			input,
+			isLoading,
+			setInput,
+			setIsLoading,
+			messages,
+			setMessages,
+			getToken,
+			solution,
+		});
 	};
 
 	const handleFeedbackClick = async () => {
@@ -204,9 +116,7 @@ export default function InterviewChatDynamic() {
 	};
 
 	useEffect(() => {
-		console.log(feedback);
 		if (feedback) {
-			console.log("sending feedback");
 			const sendFeedback = async () => {
 				const token = await getToken();
 				await fetch(
@@ -271,7 +181,7 @@ export default function InterviewChatDynamic() {
 							/>
 							<Pressable
 								style={styles.buttonWrapper}
-								onPress={handleSubmit}
+								onPress={onSubmit}
 								disabled={isLoading || disabled}
 							>
 								<LinearGradient
